fixed navigation after login todo: test the login screen test Validation.kt
