--- app/src/main/java/com/asma/tasky/authentication/data/remote/request/LoginRequest.kt
@@ -1,6 +1,0 @@
-package com.asma.tasky.authentication.data.remote.request
-
-data class LoginRequest(
-    val email:String,
-    val password:String
-)
