--- app/src/main/java/com/asma/tasky/authentication/data/remote/request/RegistrationRequest.kt
@@ -1,7 +1,0 @@
-package com.asma.tasky.authentication.data.remote.request
-
-data class RegistrationRequest(
-    val fullName:String,
-    val email:String,
-    val password:String
-)
