--- app/src/main/java/com/asma/tasky/authentication/data/remote/response/AuthenticationResponse.kt
@@ -1,7 +1,0 @@
-package com.asma.tasky.authentication.data.remote.response
-
-data class AuthenticationResponse(
-    val token:String,
-    val userId:String,
-    val fullName:String
-)
