--- app/src/main/java/com/asma/tasky/authentication/domain/util/Constants.kt
@@ -1,10 +1,0 @@
-package com.asma.tasky.authentication.domain.util
-
-object Constants {
-
-    const val MIN_NAME_LENGTH = 2
-    const val MAX_NAME_LENGTH = 50
-
-    const val MIN_PASSWORD_LENGTH = 9
-
-}
