--- app/src/main/java/com/asma/tasky/authentication/presentation/components/FormTextField.kt
@@ -1,154 +1,0 @@
-package com.asma.tasky.authentication.presentation.components
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Check
-import androidx.compose.material.icons.filled.Visibility
-import androidx.compose.material.icons.filled.VisibilityOff
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.FocusState
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.focus.onFocusChanged
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.asma.tasky.R
-import com.asma.tasky.core.presentation.ui.theme.Green
-import com.asma.tasky.core.presentation.ui.theme.LightGray
-
-
-@Composable
-fun FormTextField(
-    modifier: Modifier = Modifier,
-    text: String = "",
-    hint: String = "",
-    maxLength: Int = 400,
-    error: String = "",
-    style: TextStyle = TextStyle(
-        color = MaterialTheme.colors.onBackground
-    ),
-    backgroundColor: Color = MaterialTheme.colors.surface,
-    singleLine: Boolean = true,
-    maxLines: Int = 1,
-    keyboardType: KeyboardType = KeyboardType.Text,
-    isCheckMarkDisplayed: Boolean = false,
-    isPasswordToggleDisplayed: Boolean = keyboardType == KeyboardType.Password,
-    isPasswordVisible: Boolean = false,
-    onPasswordToggleClick: (Boolean) -> Unit = {},
-    onValueChange: (String) -> Unit
-) {
-
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-
-            .then(modifier)
-    ) {
-        TextField(
-            value = text,
-            onValueChange = {
-                if (it.length <= maxLength) {
-                    onValueChange(it)
-                }
-            },
-            maxLines = maxLines,
-            textStyle = style,
-            colors = TextFieldDefaults.textFieldColors(
-                backgroundColor = backgroundColor,
-                focusedIndicatorColor = Color.Transparent,
-                unfocusedIndicatorColor = Color.Transparent,
-                disabledIndicatorColor = Color.Transparent
-            ),
-            placeholder = {
-                Text(
-                    text = hint,
-                    style = MaterialTheme.typography.body1
-                )
-            },
-            isError = error != "",
-            keyboardOptions = KeyboardOptions(
-                keyboardType = keyboardType
-            ),
-            visualTransformation = if (!isPasswordVisible && isPasswordToggleDisplayed) {
-                PasswordVisualTransformation()
-            } else {
-                VisualTransformation.None
-            },
-            singleLine = singleLine,
-            trailingIcon = if (isPasswordToggleDisplayed) {
-                val icon: @Composable () -> Unit = {
-                    IconButton(
-                        onClick = {
-                            onPasswordToggleClick(!isPasswordVisible)
-                        }
-                    ) {
-                        Icon(
-                            imageVector = if (isPasswordVisible) {
-                                Icons.Filled.VisibilityOff
-                            } else {
-                                Icons.Filled.Visibility
-                            },
-                            tint = Color.LightGray,
-                            contentDescription = if (isPasswordVisible) {
-                                stringResource(id = R.string.visible_password_content_description)
-                            } else {
-                                stringResource(id = R.string.hidden_password_content_description)
-                            }
-                        )
-                    }
-                }
-                icon
-            } else if(isCheckMarkDisplayed){
-                val icon: @Composable () -> Unit = {
-                    IconButton(
-                        onClick = {}
-                    ) {
-                        Icon(imageVector = Icons.Filled.Check,
-                            tint = Green,
-                            contentDescription = "check mark"
-                        )
-                    }
-                }
-                icon
-            }else null,
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(
-                    color = LightGray,
-                    shape = RoundedCornerShape(10.dp)
-                ),
-        )
-        if (error.isNotEmpty()) {
-            Text(
-                text = error,
-                style = MaterialTheme.typography.body2,
-                color = MaterialTheme.colors.error,
-                textAlign = TextAlign.End,
-                modifier = Modifier
-                    .fillMaxWidth()
-            )
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun Preview() {
-    FormTextField(hint = "email", onValueChange = {}, isCheckMarkDisplayed = true)
-}
-
-
