--- app/src/main/java/com/asma/tasky/authentication/presentation/login/LoginEvent.kt
@@ -1,8 +1,0 @@
-package com.asma.tasky.authentication.presentation.login
-
-sealed class LoginEvent {
-    data class EmailEntered(val email: String): LoginEvent()
-    data class PasswordEntered(val password: String): LoginEvent()
-    object Login: LoginEvent()
-    object TogglePasswordVisibility: LoginEvent()
-}
