--- app/src/main/java/com/asma/tasky/authentication/presentation/login/LoginState.kt
@@ -1,6 +1,0 @@
-package com.asma.tasky.authentication.presentation.login
-
-data class LoginState(
-    val isLoading: Boolean = false,
-    val isPasswordVisible: Boolean = false
-)
