--- app/src/main/java/com/asma/tasky/authentication/presentation/login/LoginViewModel.kt
@@ -1,89 +1,0 @@
-package com.asma.tasky.authentication.presentation.login
-
-import androidx.compose.runtime.State
-import androidx.compose.runtime.mutableStateOf
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.asma.tasky.authentication.domain.use_case.LoginUseCase
-import com.asma.tasky.common.Resource
-import com.asma.tasky.core.domain.states.TextFieldState
-import com.asma.tasky.core.presentation.util.UiEvent
-import com.asma.tasky.core.util.UiText
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.asSharedFlow
-import kotlinx.coroutines.launch
-import javax.inject.Inject
-
-@HiltViewModel
-class LoginViewModel @Inject constructor(
-    private val loginUseCase: LoginUseCase
-) : ViewModel() {
-
-    private val _emailState = mutableStateOf(TextFieldState())
-    val emailState: State<TextFieldState> = _emailState
-
-    private val _passwordState = mutableStateOf(TextFieldState())
-    val passwordState: State<TextFieldState> = _passwordState
-
-    private val _loginState = mutableStateOf(LoginState())
-    val loginState: State<LoginState> = _loginState
-
-    private val _eventFlow = MutableSharedFlow<UiEvent>()
-    val eventFlow = _eventFlow.asSharedFlow()
-
-    fun onEvent(event: LoginEvent) {
-
-        when (event) {
-            is LoginEvent.EmailEntered -> {
-                _emailState.value = emailState.value.copy(
-                    text = event.email
-                )
-            }
-            is LoginEvent.PasswordEntered -> {
-                _passwordState.value = passwordState.value.copy(
-                    text = event.password
-                )
-            }
-            is LoginEvent.TogglePasswordVisibility -> {
-                _loginState.value = _loginState.value.copy(
-                    isPasswordVisible = !_loginState.value.isPasswordVisible
-                )
-            }
-            is LoginEvent.Login -> {
-                viewModelScope.launch {
-                    _loginState.value = _loginState.value.copy(isLoading = true)
-                    val loginResult = loginUseCase(
-                        email = emailState.value.text,
-                        password = passwordState.value.text
-                    )
-                    _loginState.value = _loginState.value.copy(isLoading = false)
-                    if (loginResult.emailError != null) {
-                        _emailState.value = emailState.value.copy(
-                            error = loginResult.emailError
-                        )
-                    }
-                    if (loginResult.passwordError != null) {
-                        _passwordState.value = _passwordState.value.copy(
-                            error = loginResult.passwordError
-                        )
-                    }
-                    when (loginResult.result) {
-                        is Resource.Success -> {
-                            _eventFlow.emit(UiEvent.OnLogin)
-                        }
-                        is Resource.Error -> {
-                            _eventFlow.emit(
-                                UiEvent.ShowSnackbar(
-                                    loginResult.result.uiText ?: UiText.unknownError()
-                                )
-                            )
-                        }
-                        else -> {}
-                    }
-
-                }
-            }
-        }
-    }
-}
