--- app/src/main/java/com/asma/tasky/authentication/presentation/register/RegisterEvent.kt
@@ -1,9 +1,0 @@
-package com.asma.tasky.authentication.presentation.register
-
-sealed class RegisterEvent {
-    data class NameEntered(val name: String): RegisterEvent()
-    data class EmailEntered(val email: String): RegisterEvent()
-    data class PasswordEntered(val password: String): RegisterEvent()
-    object Register: RegisterEvent()
-    object TogglePasswordVisibility: RegisterEvent()
-}
