--- app/src/main/java/com/asma/tasky/authentication/presentation/register/RegisterState.kt
@@ -1,9 +1,0 @@
-package com.asma.tasky.authentication.presentation.register
-
-data class RegisterState(
-    val isLoading: Boolean = false,
-    val isPasswordVisible: Boolean = false,
-    val isPasswordValid: Boolean? = null,
-    val isEmailValid: Boolean? = null,
-    val isNameValid: Boolean? = null
-)
