--- app/src/main/java/com/asma/tasky/common/Resource.kt
@@ -1,8 +1,0 @@
-package com.asma.tasky.common
-
-import com.asma.tasky.core.util.UiText
-
-sealed class Resource<T>(val data: T? = null, val uiText: UiText? = null) {
-    class Success<T>(data: T) : Resource<T>(data)
-    class Error<T>(message: UiText, data: T? = null) : Resource<T>(data, message)
-}
