--- app/src/main/java/com/asma/tasky/common/Response.kt
@@ -1,7 +1,0 @@
-package com.asma.tasky.common
-
-data class Response<T>(
-    val successful: Boolean,
-    val message: String? = null,
-    val data: T? = null
-)
