+++ app/src/main/java/com/asma/tasky/core/presentation/MainViewModel.kt
@@ -2,12 +2,18 @@ package com.asma.tasky.core.presentation
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import kotlinx.coroutines.delay
+import com.asma.tasky.feature_authentication.domain.repository.AuthenticationRepository
+import com.asma.tasky.feature_authentication.domain.util.AuthResult
+import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
+import javax.inject.Inject
 
-class MainViewModel : ViewModel() {
+@HiltViewModel
+class MainViewModel @Inject constructor(
+    private val repository: AuthenticationRepository
+) : ViewModel() {
 
     private val _isAuthenticating = MutableStateFlow(true)
     val isAuthenticating = _isAuthenticating.asStateFlow()
@@ -17,7 +23,12 @@ class MainViewModel : ViewModel() {
 
     init {
         viewModelScope.launch {
-            delay(500)
+            _isAuthenticating.value = true
+            when (repository.authenticate()) {
+                is AuthResult.Success -> _isLoggedIn.value = true
+                is AuthResult.Error -> _isLoggedIn.value = true
+                is AuthResult.Unauthorized -> _isLoggedIn.value = false
+            }
             _isAuthenticating.value = false
         }
     }
