+++ app/src/main/java/com/asma/tasky/core/presentation/components/Navigation.kt
@@ -1,19 +1,24 @@
 package com.asma.tasky.core.presentation.components
 
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.ScaffoldState
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.asma.tasky.authentication.presentation.login.LoginScreen
-import com.asma.tasky.authentication.presentation.register.RegisterScreen
+import com.asma.tasky.feature_management.agenda.AgendaScreen
+import com.asma.tasky.feature_authentication.presentation.login.LoginScreen
+import com.asma.tasky.feature_authentication.presentation.register.RegisterScreen
 import com.asma.tasky.core.util.Screen
 
+@ExperimentalComposeUiApi
 @Composable
 fun Navigation(
     navController: NavHostController,
-    startDestination: String
+    startDestination: String,
+    scaffoldState: ScaffoldState,
 ) {
     NavHost(
         navController = navController,
@@ -23,14 +28,21 @@ fun Navigation(
 
         composable(Screen.LoginScreen.route) {
             LoginScreen(
-                onNavigate = { navController.navigate(route = it) }
+                onNavigate = { navController.navigate(route = it) },
+                onLogin = {
+                    navController.popBackStack()
+                    navController.navigate(route = Screen.AgendaScreen.route)
+                          },
+                scaffoldState = scaffoldState
             )
         }
         composable(Screen.RegisterScreen.route) {
-            RegisterScreen()
+            RegisterScreen(onNavigate = {
+                navController.popBackStack()
+            }, scaffoldState = scaffoldState)
         }
         composable(Screen.AgendaScreen.route) {
-
+            AgendaScreen()
         }
     }
 }
