+++ app/src/main/java/com/asma/tasky/core/presentation/util/UiEvent.kt
@@ -6,6 +6,5 @@ import com.asma.tasky.core.util.UiText
 sealed class UiEvent : Event() {
     data class ShowSnackbar(val uiText: UiText) : UiEvent()
     data class Navigate(val route: String) : UiEvent()
-    object NavigateUp : UiEvent()
     object OnLogin : UiEvent()
 }
