+++ app/src/main/java/com/asma/tasky/core/util/Constants.kt
@@ -1,0 +1,14 @@
+package com.asma.tasky.core.util
+
+object Constants {
+
+    const val MIN_NAME_LENGTH = 2
+    const val MAX_NAME_LENGTH = 50
+
+    const val MIN_PASSWORD_LENGTH = 9
+
+    const val KEY_JWT_TOKEN = "jwt_token"
+    const val KEY_USER_ID = "userId"
+
+    const val SHARED_PREF_NAME = "shared_pref"
+}
