+++ app/src/main/java/com/asma/tasky/di/AppModule.kt
@@ -1,6 +1,10 @@
 package com.asma.tasky.di
 
-import com.asma.tasky.authentication.data.remote.util.ApiKeyInterceptor
+import android.app.Application
+import android.content.Context
+import android.content.SharedPreferences
+import com.asma.tasky.feature_authentication.data.remote.util.ApiKeyInterceptor
+import com.asma.tasky.core.util.Constants
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -14,9 +18,25 @@ import javax.inject.Singleton
 object AppModule {
     @Provides
     @Singleton
-    fun provideOkHttpClient(): OkHttpClient {
+    fun provideSharedPref(app: Application): SharedPreferences {
+        return app.getSharedPreferences(
+            Constants.SHARED_PREF_NAME,
+            Context.MODE_PRIVATE
+        )
+    }
+
+    @Provides
+    @Singleton
+    fun provideOkHttpClient(sharedPreferences: SharedPreferences): OkHttpClient {
         return OkHttpClient.Builder()
             .addInterceptor(ApiKeyInterceptor())
+            .addInterceptor {
+                val token = sharedPreferences.getString(Constants.KEY_JWT_TOKEN, "")
+                val modifiedRequest = it.request().newBuilder()
+                    .addHeader("Authorization", "Bearer $token")
+                    .build()
+                it.proceed(modifiedRequest)
+            }
             .addInterceptor(
                 HttpLoggingInterceptor().apply {
                     level = HttpLoggingInterceptor.Level.BODY
