+++ app/src/main/java/com/asma/tasky/feature_authentication/data/remote/request/LoginRequest.kt
@@ -1,0 +1,6 @@
+package com.asma.tasky.feature_authentication.data.remote.request
+
+data class LoginRequest(
+    val email:String,
+    val password:String
+)
