+++ app/src/main/java/com/asma/tasky/feature_authentication/data/remote/request/RegistrationRequest.kt
@@ -1,0 +1,7 @@
+package com.asma.tasky.feature_authentication.data.remote.request
+
+data class RegistrationRequest(
+    val fullName:String,
+    val email:String,
+    val password:String
+)
