+++ app/src/main/java/com/asma/tasky/feature_authentication/data/remote/response/AuthenticationResponse.kt
@@ -1,0 +1,7 @@
+package com.asma.tasky.feature_authentication.data.remote.response
+
+data class AuthenticationResponse(
+    val token:String,
+    val userId:String,
+    val fullName:String
+)
