+++ app/src/main/java/com/asma/tasky/feature_authentication/domain/model/LoginResult.kt
@@ -1,0 +1,10 @@
+package com.asma.tasky.feature_authentication.domain.model
+
+import com.asma.tasky.feature_authentication.domain.util.AuthError
+import com.asma.tasky.core.util.Resource
+
+data class LoginResult(
+    val emailError: AuthError? = null,
+    val passwordError: AuthError? = null,
+    val result: Resource<Unit>? = null
+)
