+++ app/src/main/java/com/asma/tasky/feature_authentication/domain/util/AuthResult.kt
@@ -1,0 +1,7 @@
+package com.asma.tasky.feature_authentication.domain.util
+
+sealed class AuthResult : com.asma.tasky.core.util.Error() {
+    object Success : AuthResult()
+    object Unauthorized : AuthResult()
+    object Error : AuthResult()
+}
