+++ app/src/main/java/com/asma/tasky/feature_authentication/presentation/login/LoginState.kt
@@ -1,0 +1,6 @@
+package com.asma.tasky.feature_authentication.presentation.login
+
+data class LoginState(
+    val isLoading: Boolean = false,
+    val isPasswordVisible: Boolean = false
+)
