+++ app/src/main/java/com/asma/tasky/feature_authentication/presentation/register/RegisterState.kt
@@ -1,0 +1,9 @@
+package com.asma.tasky.feature_authentication.presentation.register
+
+data class RegisterState(
+    val isLoading: Boolean = false,
+    val isPasswordVisible: Boolean = false,
+    val isPasswordValid: Boolean? = null,
+    val isEmailValid: Boolean? = null,
+    val isNameValid: Boolean? = null
+)
