+++ app/src/main/java/com/asma/tasky/feature_management/agenda/AgendaScreen.kt
@@ -1,0 +1,9 @@
+package com.asma.tasky.feature_management.agenda
+
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun AgendaScreen() {
+    Text(text = "Agenda Screen")
+}
